+++ LightBlue/IAzureQueueMessage.cs
@@ -1,0 +1,9 @@
+﻿namespace LightBlue
+{
+    public interface IAzureQueueMessage
+    {
+        byte[] AsBytes { get; }
+        string AsString { get; }
+        int DequeueCount { get; }
+    }
+}