+++ ReleaseNotes.md
@@ -13,6 +13,9 @@ Version 1.1.16
 
 Version 1.1.15
 --------------------
+### Features
+* Support for Azure Queue Storage
+
 ###Implementation
 * Update to NewtonSoft.Json 6.0.8